Add a own style class
I add the own style because I can pass arguments to the class with them I can control the output of the document
much better. For example during writing I compile the document more often so I like to have a quick result. The final compilation should look very good so I like to use „microtype“ to refine the output. This takes time.

Now I can pass a variable „drafting“ to the package to enable or disable the microtype package.